requirements.txt
torch
torchvision
tensorflow-gpu
lmdb
pillow
six
